+++ packages/admin/package.json
@@ -19,8 +19,9 @@
     "ra-data-json-server": "^5.7.0",
     "ra-data-simple-rest": "^5.7.0",
     "react": "^19.0.0",
-    "react-admin": "^5.7.0",
-    "react-dom": "^19.0.0"
+    "react-admin": "^5.8.0",
+    "react-dom": "^19.0.0",
+    "react-router-dom": "^7.6.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.21.0",
