+++ packages/admin/src/App.tsx
@@ -7,40 +7,26 @@ import { TasksLists } from './components/tasks/TasksLists.component';
 import { SetsList } from './components/sets/SetsList.component';
 import { StudyMappingList } from './components/studymapping/StudyMappingList.component';
 import { CasdoorProvider } from './contexts/Casdoor.context';
-import { AuthProvider, useAuth } from './contexts/Auth.context';
+import { BrowserRouter } from 'react-router-dom';
+import { authProvider } from './auth';
 
 function App() {
+  const dataSource = simpleRestProvider(config.backendURL);
+
   return (
-    <ClientProvider>
-      <CasdoorProvider>
-        <AuthProvider>
-          <InnerWrapper />
-        </AuthProvider>
-      </CasdoorProvider>
-    </ClientProvider>
+    <BrowserRouter>
+      <ClientProvider>
+        <CasdoorProvider>
+          <Admin dataProvider={dataSource} authProvider={authProvider} loginPage={false}>
+            <Resource name="sets" list={SetsList} />
+            <Resource name="tasks" list={TasksLists} />
+            <Resource name="taskCompletions" list={TaskCompletionsList} />
+            <Resource name="studymapping" list={StudyMappingList} />
+          </Admin>
+        </CasdoorProvider>
+      </ClientProvider>
+    </BrowserRouter>
   );
 }
-
-/**
- * Inner container for React admin content so the outer providers
- * can be leveraged
- */
-const InnerWrapper: React.FC = () => {
-  const dataSource = simpleRestProvider(config.backendURL);
-  const auth = useAuth();
-
-  return (
-    <>
-      {auth.authProvider &&
-        <Admin dataProvider={dataSource} authProvider={auth.authProvider}>
-          <Resource name="sets" list={SetsList} />
-          <Resource name="tasks" list={TasksLists} />
-          <Resource name="taskCompletions" list={TaskCompletionsList} />
-          <Resource name="studymapping" list={StudyMappingList} />
-        </Admin>
-      }
-    </>
-  );
-};
 
 export default App;
